fix(media-fs): check that temporary file exists before deleting it

+++ modules/Media/Entities/Image.php
@@ -147,8 +147,10 @@ class Image extends BaseMedia
             $this->fileManager
                 ->save($newImage, $this->{$name . '_key'});
 
-            // delete temporary file
-            unlink($tempFilePath);
+            if (is_file($tempFilePath)) {
+                // delete temporary file
+                unlink($tempFilePath);
+            }
         }
 
         if ($tempImagePath !== '') {
